Move serde_cbor wrappers to utilties crate
Allows reuse of functions in new test files.
