+++ Cargo.toml
@@ -22,3 +22,4 @@ syn = "1.0"
 heapless = { version = "0.5.6", default-features = false, features = ["serde"] }
 serde = { version = "1", default-features = false }
 serde_cbor = { version = "0.11.0", default-features = false }
+utilities = { path = "utilities" }
