+++ utilities/Cargo.toml
@@ -1,0 +1,10 @@
+[package]
+name = "utilities"
+version = "0.1.0"
+edition = "2021"
+
+# See more keys and their definitions at https://doc.rust-lang.org/cargo/reference/manifest.html
+
+[dependencies]
+serde = { version = "1", default-features = false }
+serde_cbor = { version = "0.11.0", default-features = false }
